refactor: 优化用户服务数据层操作

+++ src/modules/system/service/impl/SysUserServiceImpl.ts
@@ -18,6 +18,7 @@ import { SysUserRoleRepositoryImpl } from '../../repository/impl/SysUserRoleRepo
 import { ISysUserService } from '../ISysUserService';
 import { SysConfigServiceImpl } from './SysConfigServiceImpl';
 import { SysDictDataServiceImpl } from './SysDictDataServiceImpl';
+import { ADMIN_ROLE_ID } from '../../../../framework/constants/AdminConstants';
 
 /**
  * 用户 业务层处理
@@ -76,7 +77,11 @@ export class SysUserServiceImpl implements ISysUserService {
   }
   async selectUserById(userId: string): Promise<SysUser> {
     if (!userId) return null;
-    return await this.sysUserRepository.selectUserById(userId);
+    const users = await this.sysUserRepository.selectUserById([userId]);
+    if (users.length > 0) {
+      return users[0];
+    }
+    return null;
   }
 
   async selectUserRoleGroup(userName: string): Promise<string[]> {
@@ -90,7 +95,7 @@ export class SysUserServiceImpl implements ISysUserService {
   }
 
   async selectUserPostGroup(userName: string): Promise<string[]> {
-    const sysPosts = await this.sysPostRepository.selectPostsByUserName(
+    const sysPosts = await this.sysPostRepository.selectPostListByUserId(
       userName
     );
     if (sysPosts && sysPosts.length > 0) {
@@ -109,7 +114,7 @@ export class SysUserServiceImpl implements ISysUserService {
     if (uniqueId === userId) {
       return true;
     }
-    return !userId;
+    return !uniqueId;
   }
 
   async checkUniquePhone(
@@ -122,7 +127,7 @@ export class SysUserServiceImpl implements ISysUserService {
     if (uniqueId === userId) {
       return true;
     }
-    return !userId;
+    return !uniqueId;
   }
 
   async checkUniqueEmail(email: string, userId: string = ''): Promise<boolean> {
@@ -132,16 +137,16 @@ export class SysUserServiceImpl implements ISysUserService {
     if (uniqueId === userId) {
       return true;
     }
-    return !userId;
+    return !uniqueId;
   }
 
   async insertUser(sysUser: SysUser): Promise<string> {
     // 新增用户信息
     const insertId = await this.sysUserRepository.insertUser(sysUser);
     if (insertId) {
-      // 新增用户与角色管理
+      // 新增用户角色信息
       await this.insertUserRole(insertId, sysUser.roleIds);
-      // 新增用户与岗位管理
+      // 新增用户岗位信息
       await this.insertUserPost(insertId, sysUser.postIds);
     }
     return insertId;
@@ -155,11 +160,11 @@ export class SysUserServiceImpl implements ISysUserService {
     const { userId, roleIds, postIds } = sysUser;
     // 删除用户与角色关联
     await this.sysUserRoleRepository.deleteUserRole([userId]);
-    // 新增用户与角色管理
+    // 新增用户角色信息
     await this.insertUserRole(userId, roleIds);
     // 删除用户与岗位关联
     await this.sysUserPostRepository.deleteUserPost([userId]);
-    // 新增用户与岗位管理
+    // 新增用户岗位信息
     await this.insertUserPost(userId, postIds);
     return await this.sysUserRepository.updateUser(sysUser);
   }
@@ -174,11 +179,14 @@ export class SysUserServiceImpl implements ISysUserService {
     roleIds: string[] = []
   ): Promise<number> {
     if (roleIds && roleIds.length <= 0) return 0;
-    const sysUserRoles: SysUserRole[] = roleIds.map(roleId => {
-      if (roleId) {
-        return new SysUserRole(userId, roleId);
+    const sysUserRoles: SysUserRole[] = [];
+    for (const roleId of roleIds) {
+      // 管理员角色禁止操作，只能通过配置指定用户ID分配
+      if (!roleId || roleId === ADMIN_ROLE_ID) {
+        continue;
       }
-    });
+      sysUserRoles.push(new SysUserRole(userId, roleId));
+    }
     if (sysUserRoles.length <= 0) return 0;
     return await this.sysUserRoleRepository.batchUserRole(sysUserRoles);
   }
@@ -193,11 +201,13 @@ export class SysUserServiceImpl implements ISysUserService {
     postIds: string[] = []
   ): Promise<number> {
     if (postIds && postIds.length <= 0) return 0;
-    const sysUserPosts: SysUserPost[] = postIds.map(postId => {
-      if (postId) {
-        return new SysUserPost(userId, postId);
+    const sysUserPosts: SysUserPost[] = [];
+    for (const postId of postIds) {
+      if (!postId) {
+        continue;
       }
-    });
+      sysUserPosts.push(new SysUserPost(userId, postId));
+    }
     if (sysUserPosts.length <= 0) return 0;
     return await this.sysUserPostRepository.batchUserPost(sysUserPosts);
   }
@@ -208,20 +218,21 @@ export class SysUserServiceImpl implements ISysUserService {
   }
 
   async deleteUserByIds(userIds: string[]): Promise<number> {
-    // 遍历检查是否都存在
-    for (const userId of userIds) {
-      // 检查是否存在
-      const user = await this.sysUserRepository.selectUserById(userId);
-      if (!user) {
-        throw new Error('没有权限访问用户数据！');
-      }
+    // 检查是否存在
+    const users = await this.sysUserRepository.selectUserById(userIds);
+    if (users.length <= 0) {
+      throw new Error('没有权限访问用户数据！');
     }
-
-    // 删除用户与角色关联
-    await this.sysUserRoleRepository.deleteUserRole(userIds);
-    // 删除用户与岗位关联
-    await this.sysUserPostRepository.deleteUserPost(userIds);
-    return await this.sysUserRepository.deleteUserByIds(userIds);
+    if (users.length === userIds.length) {
+      // 删除用户与角色关联
+      await this.sysUserRoleRepository.deleteUserRole(userIds);
+      // 删除用户与岗位关联
+      await this.sysUserPostRepository.deleteUserPost(userIds);
+      // ... 注意其他userId进行关联的表
+      // 删除用户
+      return await this.sysUserRepository.deleteUserByIds(userIds);
+    }
+    return 0;
   }
 
   async importUser(
@@ -271,7 +282,9 @@ export class SysUserServiceImpl implements ISysUserService {
 
       if (newSysUser.phonenumber) {
         if (validMobile(newSysUser.phonenumber)) {
-          const uniquePhone = await this.checkUniquePhone(newSysUser.phonenumber);
+          const uniquePhone = await this.checkUniquePhone(
+            newSysUser.phonenumber
+          );
           if (!uniquePhone) {
             failureNum++;
             failureMsgArr.push(
